Still currently working on it, halfway commit
Powerup quadrants fixed (was giving a crash error on random numbers sometimes)
Still need to fix that up a bit so that they show up on screen now since the quadrants are correct (they're all pretty much offscreen at the moment or do not last long enough)

Explosions are in, they do not remove themselves for now.  Will be doing this in a bit.

Fixed iPad background size

Announcement between waves is in, currently waves are 0.1s for debugging purposes